Moved ingester clients pool to the distributor pkg
