+++ pkg/distributor/ingester_client_pool.go
@@ -1,4 +1,4 @@
-package client
+package distributor
 
 import (
 	"flag"
